backend/app/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9 
pydantic==2.9.2 
python-dotenv==1.0.1 
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
email-validator==2.1.1
